MMT-3410: Rename directory 'CustomMultiSelect' to 'CustomMultiSelectWidget'
