+++ static/src/js/schemas/uiSchemas/tools/organization.js
@@ -1,4 +1,4 @@
-import CustomMultiSelectWidget from '../../../components/CustomMultiSelect/CustomMultiSelectWidget'
+import CustomMultiSelectWidget from '../../../components/CustomMultiSelectWidget/CustomMultiSelectWidget'
 import CustomTextWidget from '../../../components/CustomTextWidget/CustomTextWidget'
 
 const organizationUiSchema = {
